requirements.txt
about-time==4.2.1
alive-progress==3.1.5
attrs==23.2.0
autograd==1.6.2
blinker==1.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cma==3.2.2
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
Deprecated==1.2.14
dill==0.3.8
filelock==3.14.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
future==1.0.0
grapheme==0.6.0
humanfriendly==10.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jstyleson==0.0.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mpmath==1.3.0
natsort==8.4.0
ncnn==1.0.20240410
networkx==3.3
ninja==1.11.1.1
nncf==2.10.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
onnx==1.16.0
onnxruntime==1.17.3
onnxsim==0.4.36
opencv-python==4.9.0.80
openvino==2024.1.0
openvino-telemetry==2024.1.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
portalocker==2.8.2
protobuf==5.26.1
psutil==5.9.8
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pydot==2.0.0
Pygments==2.17.2
pymoo==0.6.1.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
sympy==1.12
tabulate==0.9.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
triton==2.3.0
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.7
urllib3==2.2.1
Werkzeug==3.0.2
wrapt==1.16.0
